# apps/api/openapi/v1.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuthHub Public API",
    "version": "1.0.0",
    "description": "Versioned public contract for agency developers. Additive-only change within a version; breaking change ships as a new version."
  },
  "servers": [
    {
      "url": "https://api.authhub.example/api/v1"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped agency API key as `Bearer <secret>`. The required scope is listed per operation; failures name the missing scope."
      }
    }
  },
  "paths": {
...
    "/webhook-event-types": {
      "get": {
        "summary": "List the subscribable webhook event taxonomy with descriptions.",
        "security": [
          {
            "ApiKeyAuth": [
              "webhooks:read"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                }
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "webhook.test",
                              "access_request.partial",
                              "access_request.completed",
                              "access_request.revoked",
                              "access_request.expired",
                              "connection.status_changed"
                            ]
                          },
                          "description": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "type",
                          "description"
                        ]
                      }
                    },
                    "error": {
                      "type": "null",
                      "nullable": true
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "INVALID_API_KEY"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Missing required scope",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "MISSING_SCOPE",
                            "TIER_ACCESS_DENIED"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded. Please try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "RATE_LIMIT_EXCEEDED"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Unable to verify API access for this plan. Please try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "TIER_CHECK_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      },
                      "required": [
                        "requestId"
                      ],
                      "additionalProperties": true
                    }
                  },
                  "required": [
                    "data",
                    "error",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}